Continuous training for FCP & PP

+++ MARL/FCP_train.py
@@ -12,7 +12,7 @@ from examples.rllib import utils
 from meltingpot.python import substrate
 
 def get_config(
-    substrate_name: str = "chicken_in_the_matrix__repeated_small",
+    substrate_name: str = "stag_hunt_in_the_matrix__repeated_medium",
     num_rollout_workers: int = 2,
     rollout_fragment_length: int = 100,
     train_batch_size: int = 1600,
@@ -123,42 +123,65 @@ def main():
   tune.register_env("meltingpot", utils.env_creator)
 
   # parameters
-  save_path = './MARL/FCP_logs/c3s'
-  checkpoints_path = './MARL/FCP_checkpoints/c3s'
-  log_path = './MARL/FCP_outputs/c3s.txt'
+  save_path = './MARL/FCP_logs/sh3m'
+  checkpoints_path = './MARL/FCP_checkpoints/sh3m'
+  log_path = './MARL/FCP_outputs/sh3m.txt'
 
-  checkpoints_pool = './MARL/SP_checkpoints/c5s'
+  checkpoints_pool = './MARL/SP_checkpoints/sh5m'
 
-  checkpoint_freq = 1
-  num_gens = 100
+  checkpoint_freq = 125
+  num_gens = 25
   seeds = [11,22,33]
+
+  continuous_training = False
+  starting_gen = 0
 
   gen_len = checkpoint_freq * config.train_batch_size
   num_seeds = len(seeds)
 
   # clear output
-  f = open(log_path, "w")
-  f.close()
+  if not continuous_training:
+    f = open(log_path, "w")
+    f.close()
 
   # Initialize ray, train and save
   ray.init()
 
   # logging
-  timesteps = [[] for j in range(num_seeds)]
-  policy_reward_min = [[[] for i in range(2)] for j in range(num_seeds)]
-  policy_reward_mean = [[[] for i in range(2)] for j in range(num_seeds)]
-  policy_reward_max = [[[] for i in range(2)] for j in range(num_seeds)]
+  if not continuous_training:
+    timesteps = [[] for j in range(num_seeds)]
+    policy_reward_min = [[[] for i in range(2)] for j in range(num_seeds)]
+    policy_reward_mean = [[[] for i in range(2)] for j in range(num_seeds)]
+    policy_reward_max = [[[] for i in range(2)] for j in range(num_seeds)]
+  else:
+    raw_data = np.load(save_path+'.npz')
+    timesteps = raw_data['timesteps'].tolist()
+    policy_reward_min = raw_data['policy_reward_min'].tolist()
+    policy_reward_mean = raw_data['policy_reward_mean'].tolist()
+    policy_reward_max = raw_data['policy_reward_max'].tolist()
 
   # A dictionary of lists of checkpoints for each seed/population
 
-  for f in os.scandir(checkpoints_path):
-    if f.is_dir():
-      shutil.rmtree(f)
-  checkpoints_dict = {}
-  for seed in range(num_seeds):
-    checkpoints_dict[f'Seed_{seed}'] = []
+  if not continuous_training:
+    for f in os.scandir(checkpoints_path):
+      if f.is_dir():
+        shutil.rmtree(f)
+    checkpoints_dict = {}
+    for seed in range(num_seeds):
+      checkpoints_dict[f'Seed_{seed}'] = []
+  else:
+    checkpoints_dict = {}
+    for seed in range(num_seeds):
+      current_seed_path = os.path.join(checkpoints_path, f'seed_{seed}')
+      dir_list = sorted(os.listdir(current_seed_path))
+      latest_gen_path = os.path.join(current_seed_path, dir_list[-1])
+      latest_checkpoint_path = os.path.join(latest_gen_path, os.listdir(latest_gen_path)[-1])
+      checkpoints_dict[f'Seed_{seed}'] = [latest_checkpoint_path]
 
   for gen in range(num_gens):
+    if continuous_training:
+      if gen <= starting_gen:
+        continue
     for seed in range(num_seeds):
       # config.env_config["seed"] = seeds[seed]
       ppo = PPO(config=config.to_dict())
@@ -176,7 +199,7 @@ def main():
 
       # Set Player 1's policy
       if (len(checkpoints_dict[f"Seed_{seed}"]) > 0): # random choose from pool after 1st step
-        opp_seed = random.choice(os.listdir(checkpoints_pool))
+        opp_seed = random.choice(os.listdir(checkpoints_pool)[:num_seeds])
         opp_gen = random.choice(os.listdir(os.path.join(checkpoints_pool, opp_seed)))
         opp_checkpoint = random.choice(os.listdir(os.path.join(checkpoints_pool, opp_seed, opp_gen)))
         opp_checkpoint_dir = os.path.join(checkpoints_pool, opp_seed, opp_gen, opp_checkpoint)
